make cielab_util and finalise
